93: Refactor directory name (#97)
