+++ src/uploaded-files/contracts/uploaded-file-interface.ts
